+++ src/entities/comment/ui/index.ts
@@ -1,0 +1,1 @@
+export { CommentList } from "./CommentList"
